+++ tests/integ/modin/test_faster_pandas.py
@@ -520,6 +520,124 @@ def test_expanding(session, func):
         assert_frame_equal(snow_result, native_result, check_dtype=False)
 
 
+@pytest.mark.parametrize("property_name", ["groups", "indices"])
+@sql_count_checker(query_count=3)
+def test_groupby_properties(session, property_name):
+    with session_parameter_override(
+        session, "dummy_row_pos_optimization_enabled", True
+    ):
+        # create tables
+        table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
+        session.create_dataframe(
+            native_pd.DataFrame([[2, 12], [2, 11], [3, 13]], columns=["A", "B"])
+        ).write.save_as_table(table_name, table_type="temp")
+
+        # create snow dataframes
+        df = pd.read_snowflake(table_name).sort_values("B", ignore_index=True)
+        snow_result = getattr(df.groupby("A"), property_name)
+
+        # verify that the input dataframe has a populated relaxed query compiler
+        assert df._query_compiler._relaxed_query_compiler is not None
+        assert df._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+
+        # create pandas dataframes
+        native_df = df.to_pandas()
+        native_result = getattr(native_df.groupby("A"), property_name)
+
+        # compare results
+        snow_result = {k: list(v) for k, v in snow_result.items()}
+        native_result = {k: list(v) for k, v in native_result.items()}
+        assert snow_result == native_result
+
+
+@pytest.mark.parametrize(
+    "func",
+    [
+        "first",
+        "last",
+        "rank",
+        "shift",
+        "cumcount",
+        "cumsum",
+        "cummin",
+        "cummax",
+        "any",
+        "all",
+        "unique",
+    ],
+)
+@sql_count_checker(query_count=3)
+def test_groupby_no_param_functions(session, func):
+    with session_parameter_override(
+        session, "dummy_row_pos_optimization_enabled", True
+    ):
+        # create tables
+        table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
+        session.create_dataframe(
+            native_pd.DataFrame([[2, 12], [2, 11], [3, 13]], columns=["A", "B"])
+        ).write.save_as_table(table_name, table_type="temp")
+
+        # create snow dataframes
+        df = pd.read_snowflake(table_name).sort_values("B", ignore_index=True)
+        snow_result = getattr(df.groupby("A")["B"], func)()
+
+        # verify that the input dataframe has a populated relaxed query compiler
+        assert df._query_compiler._relaxed_query_compiler is not None
+        assert df._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+        # verify that the output dataframe also has a populated relaxed query compiler
+        assert snow_result._query_compiler._relaxed_query_compiler is not None
+        assert (
+            snow_result._query_compiler._relaxed_query_compiler._dummy_row_pos_mode
+            is True
+        )
+
+        # create pandas dataframes
+        native_df = df.to_pandas()
+        native_result = getattr(native_df.groupby("A")["B"], func)()
+
+        # compare results
+        assert_series_equal(snow_result, native_result, check_dtype=False)
+
+
+@pytest.mark.parametrize(
+    "func, param",
+    [
+        ("get_group", "a"),
+    ],
+)
+@sql_count_checker(query_count=3)
+def test_groupby_one_param_functions(session, func, param):
+    with session_parameter_override(
+        session, "dummy_row_pos_optimization_enabled", True
+    ):
+        # create tables
+        table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
+        session.create_dataframe(
+            native_pd.DataFrame([["a", 12], ["a", None], ["b", 13]], columns=["A", "B"])
+        ).write.save_as_table(table_name, table_type="temp")
+
+        # create snow dataframes
+        df = pd.read_snowflake(table_name)
+        snow_result = getattr(df.groupby("A"), func)(param)
+
+        # verify that the input dataframe has a populated relaxed query compiler
+        assert df._query_compiler._relaxed_query_compiler is not None
+        assert df._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+        # verify that the output dataframe also has a populated relaxed query compiler
+        assert snow_result._query_compiler._relaxed_query_compiler is not None
+        assert (
+            snow_result._query_compiler._relaxed_query_compiler._dummy_row_pos_mode
+            is True
+        )
+
+        # create pandas dataframes
+        native_df = df.to_pandas()
+        native_result = getattr(native_df.groupby("A"), func)(param)
+
+        # compare results
+        assert_frame_equal(snow_result, native_result, check_dtype=False)
+
+
 @pytest.mark.parametrize(
     "func",
     [
@@ -618,45 +736,6 @@ def test_groupby_apply(session):
         assert_frame_equal(
             snow_result, native_result, check_dtype=False, check_index_type=False
         )
-
-
-@pytest.mark.parametrize(
-    "func, param",
-    [
-        ("get_group", "a"),
-    ],
-)
-@sql_count_checker(query_count=3)
-def test_groupby_one_param_functions(session, func, param):
-    with session_parameter_override(
-        session, "dummy_row_pos_optimization_enabled", True
-    ):
-        # create tables
-        table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
-        session.create_dataframe(
-            native_pd.DataFrame([["a", 12], ["a", None], ["b", 13]], columns=["A", "B"])
-        ).write.save_as_table(table_name, table_type="temp")
-
-        # create snow dataframes
-        df = pd.read_snowflake(table_name)
-        snow_result = getattr(df.groupby("A"), func)(param)
-
-        # verify that the input dataframe has a populated relaxed query compiler
-        assert df._query_compiler._relaxed_query_compiler is not None
-        assert df._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
-        # verify that the output dataframe also has a populated relaxed query compiler
-        assert snow_result._query_compiler._relaxed_query_compiler is not None
-        assert (
-            snow_result._query_compiler._relaxed_query_compiler._dummy_row_pos_mode
-            is True
-        )
-
-        # create pandas dataframes
-        native_df = df.to_pandas()
-        native_result = getattr(native_df.groupby("A"), func)(param)
-
-        # compare results
-        assert_frame_equal(snow_result, native_result, check_dtype=False)
 
 
 @sql_count_checker(query_count=9, join_count=1)